AppliedPython/ht03_FastApi_url_shortener/requirements.txt
fastapi[all]==0.115.5
fastapi-cache2[redis]==0.2.2
pydantic==2.10.6
uvicorn==0.34.0
SQLAlchemy==2.0.39
sqlmodel==0.0.24
psycopg2-binary==2.9.10
redis==4.6.0
python-jose==3.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4